Add indentation to sub-bullet points
It wasn't until I pushed that it was spotted my editor had removed a number of intentional trailing spaces. These were being used to get the content rendering such that it should as a sub-point under the main bullet point.

After discussion with @surgisonb it was agreed that this is something that is likely to happen to others working on the README, so an alternative way to achieve something similar would be great.

This change uses an additional indent to receive a similar effect.