Add LICENSE and sections to README (#20)
https://github.com/DEFRA/dst-guides/blob/master/github/new_projects.md#add-a-licence-and-readme

As per our guides for new repos, and to truly make this open for others to use, this change adds our standard license plus sections about it and contributing to the README.

Note. It wasn't until I pushed that it was spotted my editor had removed a number of intentional trailing spaces. These were being used to get the content rendering such that it should as a sub-point under the main bullet point.

After discussion with @surgisonb it was agreed that this is something that is likely to happen to others working on the README, so an alternative way to achieve something similar would be great.

So this change also applies an additional indent to receive a similar effect.